Add getWorkspaceAccessTokenWithScope which only gets token and doesn't store.  Returns expiresAt
